appseo-video-overview sayfa islemleri tamamlandi

+++ src/OGWeb.Core/Commands/Videos/DeleteVideoCommand.cs
@@ -1,3 +1,4 @@
+using MediaBalansDocument.Library;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using OGWeb.Core.Interfaces.Repositories;
@@ -10,12 +11,14 @@ public class DeleteVideoCommand:IRequest<CustomResponse<NoContent>>
     public Guid Id { get; set; }
     public class DeleteVideoCommandHandler : IRequestHandler<DeleteVideoCommand, CustomResponse<NoContent>>
     {
-         private readonly IWriteRepositoryManager _writeRepositoryManager;
+        private readonly IWriteRepositoryManager _writeRepositoryManager;
+        private readonly IReadRepositoryManager _readRepositoryManager;
         private readonly ILogger<DeleteVideoCommandHandler> _logger;
 
-        public DeleteVideoCommandHandler(IWriteRepositoryManager writeRepositoryManager, ILogger<DeleteVideoCommandHandler> logger)
+        public DeleteVideoCommandHandler(IWriteRepositoryManager writeRepositoryManager, IReadRepositoryManager readRepositoryManager, ILogger<DeleteVideoCommandHandler> logger)
         {
             _writeRepositoryManager = writeRepositoryManager;
+            _readRepositoryManager = readRepositoryManager;
             _logger = logger;
         }
 
@@ -26,7 +29,21 @@ public class DeleteVideoCommand:IRequest<CustomResponse<NoContent>>
                 _logger.LogError($"Querying for single videoId: {request.Id}");
                 throw new ArgumentNullException($"Querying for single videoId: {request.Id}");
             }
+            var result = await _readRepositoryManager.VideoRepository.GetByIdVideoAsync(request.Id);
+
+            if(result == null)
+            {
+                _logger.LogError($"Querying for single videoId: {result.Id}");
+                throw new ArgumentNullException($"Querying for single videoId: {result.Id}");
+            }
+
+            if (!string.IsNullOrEmpty(result.ImageUrl))
+            {
+                FileUploader.FolderRemove(result.ImageUrl);
+            }
+
             await _writeRepositoryManager.VideoRepository.RemoveVideoAsync(request.Id);
+
             return CustomResponse<NoContent>.Success(204);
         }
     }
